requirements.txt
faiss-cpu>=1.8.0; platform_system != "Windows"
numpy>=1.26.0
openai>=1.30.0
pypdf>=4.2.0
python-docx>=1.1.0
python-dotenv>=1.0.1
sentence-transformers>=2.7.0
streamlit>=1.35.0
